Mark possible bug around threading
When running some of the LazyCache tests for the ISessionManager,
running all the tests at once, with the same key used for all the tests
made the value get stuck on the first test case.

Which is weird, because we're new'ing up a new SUT for each
test run within the [Theory] method.  It's possible that LazyCache
introduces some shared context (somehow) which is causing tests
which use the same feature flag name to misbehave.

AFAIK, we *should* be using ConfigureAwait(false) in these places, but
it's something that I'm not completely sure of.