Reorganize packages for v1.2
- The caching IFeatureManagerSnapshot implementation is now in the base 
package.  Because 99% of the time, you're going to use it over just using 
IFeatureManager.  This means the base package now takes a dependency 
on LazyCache.

- There are enough ISessionManager implementations to warrant putting the 
ones which are .NET Standard 2.0 compatible in a top-level package.  This 
will reduce some of the code duplication that was starting to occur in the
.Core and .Framework packages.