Add ILussatiteSessionManager and SQL write-back
- Add the ILussatiteSessionManager interface which inherits from the
ISessionManager interface and adds SetNullableAsync().  This method
does not get called by Microsoft IFeatureManager implementations, 
which makes it a safer way to set a value in the database table via
some out-of-band process such as a separate user-interface.

- Add "setValueCommandFactory" and "setNullableValueCommandFactory"
arguments to the constructor of SqlSessionManager.  These DbCommand
arguments, if not null, will enable writing back to the SQL database table
using SetAsync() and SetNullableAsync().

Note:  Most implementations should prefer to define the DbCommand for
SetNullableAsync() and use that from a separate UI instead of allowing
the database value to be set during execution of the Microsoft 
IFeatureManager.IsEnabledAsync implementation.