Allow dynamic schema/table/column names in queries
Things get really messy in SQL when you try and allow the consumer
of your library to customize schema/table/column names.  The
best that can be done is implement a very restrictive regex around
identifier names and wrap most identifiers in square brackets.

It's still possible to end up with a SQL injection attack.  But with
identifiers limited to just 64 total characters, it should be difficult.

Other mitigation is that the SQL connection string should be limited
to SELECT/INSERT/UPDATE.