Separate "cached" settings from SQL settings
Having CachedSqlSessionManagerSettings inherit from SqlSessionManagerSettings
was starting to introduce difficulties in the code.  By keeping the cache SQL manager
settings separate, it simplifies a lot of code.

The trade-off is the more complicated constructor.  But you only have to
deal with that if you want a non-default cache TimeSpan value.