Remove IAppCache injection from constructors
Because of how the cache keys are constructed (just a prefix plus
the feature name), using a shared IAppCache would cause problems
in cases where you are trying to do per-user / per-session / per-request
feature flags.

This can be revisited again in the future when the cache key names
are reworked to have a unique prefix to the object instance of the
caching IFeatureManagerSnapshot or caching ISessionManager.