Make test feature names unique to the test
As a general guideline, it is useful to have unique strings for any feature
names in tests.  Not only to avoid tests stepping on each other if the
context happens to be shared, but also to make it easier to track down
which test failed.

(I also keep forgetting to change the prefix when copying code around.)