Use global application's IAppCache if possible
Rework the calculation for the cache-keys so that CachedSqlSessionManager 
can use the global LazyCache IAppCache instance.  This reduces the number of
objects created and helps with caching of GUID-specific feature values across
multiple API calls.

The provided IAppCache can be a singleton after this change.

Because of implementation details, this has to be stored on the 
SqlSessionManagerSettings object instead of CachedSqlSessionManagerSettings.